observability: remediate AV.0 boundary violations
