+++ crates/atm-daemon/src/daemon/mod.rs
@@ -3,6 +3,7 @@ pub mod dedup;
 pub mod event_loop;
 pub mod gh_monitor_router;
 pub mod log_writer;
+pub mod observability;
 pub mod pid_backend_validation;
 pub mod session_registry;
 pub mod shutdown;
