+++ crates/sc-composer/Cargo.toml
@@ -15,7 +15,6 @@ thiserror.workspace = true
 minijinja.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-sc-observability.workspace = true
 serde_yaml = "0.9"
 
 [dev-dependencies]
